🐛 fix bundler bugs by importing rollup to the project 💥 Convert project into typescript 💄 Fix some css mistakes ⊖ Remove examples, eslint, babel, prettier (provisionally) ⊖ Put utils.js inside Datepicker.tsx ⊕ Add index.ts to export component ⊕ Add tsconfig ⊕ Add rollup config ⬆️ update package.json with new setup ⬆️ update readme
