--- .eslintrc
@@ -1,5 +1,0 @@
-{
-    "extends": ["react-app"],
-    "rules": {},
-    "plugins": ["jsx-a11y"]
-  }
