--- .prettierrc
@@ -1,5 +1,0 @@
-{
-    "semi": false,
-    "singleQuote": true,
-    "tabWidth": 2
-}
