+++ package.json
@@ -1,91 +1,50 @@
 {
   "name": "basic-datepicker-react",
+  "version": "0.1.6",
   "description": "A simple and basic reusable datepicker component using react",
   "author": "thra",
-  "version": "0.1.4",
-  "private": false,
-  "main": "build/index.js",
-  "module": "build/index.js",
-  "files": [
-    "build",
-    "README.md"
-  ],
   "repository": {
     "type": "git",
     "url": "git+https://github.com/thra/basic-datepicker-react.git"
+  },
+  "private": false,
+  "main": "dist/cjs/index.js",
+  "module": "dist/esm/index.js",
+  "files": [
+    "dist"
+  ],
+  "types": "dist/index.d.ts",
+  "scripts": {
+    "rollup": "rollup -c"
+  },
+  "resolutions": {
+    "src/components/react": "18.2.0",
+    "src/components/react-dom": "18.2.0"
+  },
+  "devDependencies": {
+    "@rollup/plugin-commonjs": "^24.0.0",
+    "@rollup/plugin-node-resolve": "^15.0.1",
+    "@rollup/plugin-typescript": "^10.0.1",
+    "@types/react": "^18.0.26",
+    "@types/react-dom": "^18.0.10",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "rollup": "^3.8.1",
+    "rollup-plugin-dts": "^5.1.0",
+    "rollup-plugin-peer-deps-external": "^2.2.4",
+    "rollup-plugin-postcss": "^4.0.2",
+    "tslib": "^2.4.1",
+    "typescript": "^4.9.4",
+    "sass": "^1.57.1"
+  },
+  "peerDependencies": {
+    "react": ">=17.0.0",
+    "react-dom": ">=17.0.0"
   },
   "keywords": [
     "react",
     "components",
     "ui",
     "datepicker"
-  ],
-  "scripts": {
-    "start": "cross-env PORT=9091 react-scripts start",
-    "build-examples": "react-scripts build",
-    "test": "react-scripts test",
-    "eject": "react-scripts eject",
-    "oldBuild": "cross-env NODE_ENV=production babel --presets @babel/preset-react src/lib --out-dir build --copy-files --ignore __tests__,spec.js,test.js,__snapshots__",
-    "build": "rm -rf build && cross-env NODE_ENV=production babel src/lib --out-dir build --copy-files"
-  },
-  "eslintConfig": {
-    "extends": [
-      "react-app",
-      "react-app/jest"
-    ]
-  },
-  "browserslist": {
-    "production": [
-      ">0.2%",
-      "not dead",
-      "not op_mini all"
-    ],
-    "development": [
-      "last 1 chrome version",
-      "last 1 firefox version",
-      "last 1 safari version"
-    ]
-  },
-  "devDependencies": {
-    "@babel/cli": "^7.19.3",
-    "@babel/core": "^7.19.3",
-    "@babel/preset-env": "^7.19.3",
-    "@babel/preset-react": "^7.18.6",
-    "@rollup/plugin-commonjs": "^24.0.0",
-    "@rollup/plugin-node-resolve": "^15.0.1",
-    "@rollup/plugin-typescript": "^10.0.1",
-    "@testing-library/jest-dom": "^5.16.5",
-    "@testing-library/react": "^13.4.0",
-    "@testing-library/user-event": "^14.4.3",
-    "@types/react": "^18.0.26",
-    "@types/react-dom": "^18.0.10",
-    "cross-env": "^7.0.3",
-    "eslint": "^8.21.0",
-    "eslint-config-airbnb": "^19.0.4",
-    "eslint-plugin-import": "^2.26.0",
-    "eslint-plugin-jsx-a11y": "^6.6.1",
-    "eslint-plugin-react": "^7.30.1",
-    "eslint-plugin-react-hooks": "^4.6.0",
-    "prettier": "^2.7.1",
-    "react": "^18.2.0",
-    "react-dom": "^18.2.0",
-    "react-hook-form": "^7.41.0",
-    "react-scripts": "5.0.1",
-    "rollup": "^3.8.1",
-    "rollup-plugin-dts": "^5.1.0",
-    "rollup-plugin-peer-deps-external": "^2.2.4",
-    "rollup-plugin-postcss": "^4.0.2",
-    "sass": "^1.54.4",
-    "tslib": "^2.4.1",
-    "typescript": "^4.9.4"
-  },
-  "dependencies": {
-    "@babel/polyfill": "^7.12.1",
-    "react": "^18.2.0",
-    "react-dom": "^18.2.0"
-  },
-  "peerDependencies": {
-    "react": ">=17.0.0",
-    "react-dom": ">=17.0.0"
-  }
+  ]
 }
