+++ src/components/Datepicker/Datepicker.scss
@@ -69,10 +69,11 @@
 	}
 	
 	.calendar-header-menu-button {
-	  height: 8px;
+	  font-size: 18px;
 	  padding-right: 5px;
 	  padding-bottom: 2px;
 	  padding-left: 4px;
+	  color: var(--FONT-COLOR-strong);
 	  border: none;
 	}
