+++ src/components/Datepicker/index.ts
@@ -1,0 +1,1 @@
+export {default} from './Datepicker'
